sdlc_with_llm/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.10.6
python-dotenv==1.0.1
pytest==8.0.0
httpx==0.26.0
pytest-asyncio==0.21.0
aiosqlite==0.19.0
asyncpg==0.29.0
